Restructured project to a have a game.go
